refactoring PageObjectModelResources getTestCaseUser() to testUser()
