+++ src/test/java/com/qauber/pagestest/CreateOrganizationTestCase.java
@@ -21,7 +21,7 @@ public class CreateOrganizationTestCase extends PageObjectModelResources {
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
     }
 
 
@@ -32,7 +32,7 @@ public class CreateOrganizationTestCase extends PageObjectModelResources {
 
         Faker faker = new Faker();
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(3000);
 
         getNavBar().clickEntities();
