+++ src/test/java/com/qauber/pagestest/InviteUserFromUsersListTestCase.java
@@ -19,7 +19,7 @@ public class InviteUserFromUsersListTestCase extends PageObjectModelResources {
     public void setUp() {
         driver = new ChromeDriver();
      //   setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
 
         setUpWithUser(User.UserType.AU, driver);
     }
@@ -32,7 +32,7 @@ public class InviteUserFromUsersListTestCase extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getNavBar().usersButton().click();
