+++ src/test/java/com/qauber/pagestest/RegularUserCannotReportNotAssignedEntitiesTestCase.java
@@ -3,7 +3,6 @@ package com.qauber.pagestest;
 import com.qauber.config.Config;
 import com.qauber.pagesresource.PageObjectModelResources;
 import com.qauber.pagesresource.User;
-import com.qauber.pagesresource.UserFactory;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
@@ -29,7 +28,7 @@ public class RegularUserCannotReportNotAssignedEntitiesTestCase extends PageObje
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
 
 //        setUpWithUser(User.UserType.AU, driver);
     }
@@ -44,7 +43,7 @@ public class RegularUserCannotReportNotAssignedEntitiesTestCase extends PageObje
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getNavBar().usersButton().click();
@@ -81,7 +80,7 @@ public class RegularUserCannotReportNotAssignedEntitiesTestCase extends PageObje
         setUpWithUser(User.UserType.RU, driver);
         Thread.sleep(2000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getNavBar().addReportButton().click();
