+++ src/test/java/com/qauber/pagestest/ReportsExportReportListAsPDFTestCase.java
@@ -24,7 +24,7 @@ public class ReportsExportReportListAsPDFTestCase extends PageObjectModelResourc
     public void reportsExportReportListAsPDF() throws InterruptedException {
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
         getReports().exportReportsListButton().click();
     }
