+++ src/test/java/com/qauber/pagestest/ReportsViewReportTestCase.java
@@ -27,7 +27,7 @@ public class ReportsViewReportTestCase extends PageObjectModelResources {
         int rowindex = 2;
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
         getNavBar().clickReports();
         getReports().selectReport(rowindex).click();
