+++ src/test/java/com/qauber/pagestest/SuperAdminDeleteEntityTestCase.java
@@ -30,7 +30,7 @@ public class SuperAdminDeleteEntityTestCase extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getNavBar().entitiesButton().click();
