+++ src/test/java/com/qauber/sanity/CancelSubscriptionTestCase.java
@@ -23,7 +23,7 @@ public class CancelSubscriptionTestCase extends PageObjectModelResources {
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
     }
 
     @Test
@@ -36,7 +36,7 @@ public class CancelSubscriptionTestCase extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getHeader().userName().click();
@@ -63,7 +63,7 @@ public class CancelSubscriptionTestCase extends PageObjectModelResources {
         getSubscriptionSettings().cancelSubscriptionButton().click();
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getNavBar().entitiesButton().click();
