+++ src/test/java/com/qauber/sanity/CompanyCreationSAUTestCase.java
@@ -22,7 +22,7 @@ public class CompanyCreationSAUTestCase extends PageObjectModelResources {
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
     }
 
     @Test
