+++ src/test/java/com/qauber/sanity/RegularUserNotDeleteOrChangeSubscription.java
@@ -25,7 +25,7 @@ public class RegularUserNotDeleteOrChangeSubscription extends PageObjectModelRes
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
     }
 
     @Test
@@ -37,7 +37,7 @@ public class RegularUserNotDeleteOrChangeSubscription extends PageObjectModelRes
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getHeader().userName().click();
@@ -76,7 +76,7 @@ public class RegularUserNotDeleteOrChangeSubscription extends PageObjectModelRes
         setUpWithUser(User.UserType.RU, driver);
         Thread.sleep(2000);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
 
         getHeader().userName().click();
