+++ src/test/java/com/qauber/sanity/SearchReportsByContainsText.java
@@ -29,7 +29,7 @@ public class SearchReportsByContainsText extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(sleepTime*2);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(sleepTime*2);
     }
