+++ src/test/java/com/qauber/sanity/SearchReportsByCreatedDate.java
@@ -30,7 +30,7 @@ public class SearchReportsByCreatedDate extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(sleepTime*2);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(sleepTime*2);
     }
