+++ src/test/java/com/qauber/sanity/SearchReportsPagination.java
@@ -28,7 +28,7 @@ public class SearchReportsPagination extends PageObjectModelResources {
         driver.get(Config.getBaseURL());
         Thread.sleep(sleepTime*2);
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(sleepTime*2);
     }
