+++ src/test/java/com/qauber/sanity/UserDeletionTestCase.java
@@ -1,12 +1,10 @@
 package com.qauber.sanity;
 
-import com.github.javafaker.Faker;
 import com.qauber.config.Config;
 import com.qauber.pagesresource.PageObjectModelResources;
 import com.qauber.pagesresource.User;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -21,7 +19,7 @@ public class UserDeletionTestCase extends PageObjectModelResources {
     public void setUp() {
         driver = new ChromeDriver();
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
     }
 
 
@@ -34,7 +32,7 @@ public class UserDeletionTestCase extends PageObjectModelResources {
         //String emailName = "kennith.willms@hotmail.com";
 
 
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(3000);
 
         getNavBar().clickUsers();
