+++ src/test/java/com/qauber/sanity/UsersAssignPermissions.java
@@ -26,20 +26,20 @@ public class UsersAssignPermissions extends PageObjectModelResources {
     @Test
     public void testAsSAU() throws AWTException, InterruptedException {
         setUpWithUser(User.UserType.SAU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
         assignPermissionsUsers();
     }
     @Test(priority = 1)
     public void testAsAU() throws AWTException, InterruptedException {
         setUpWithUser(User.UserType.AU, driver); //pass userType and browser. see ~/QAUberTestConfig
-        //setUpWithUser creates TestCaseUser, access with getTestCaseUser()
+        //setUpWithUser creates TestCaseUser, access with testUser()
         assignPermissionsUsers();
     }
 
     public void assignPermissionsUsers() throws InterruptedException, AWTException {
         driver.get(Config.getBaseURL());
         Thread.sleep(2000);
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(2000);
         getNavBar().usersButton().click();
         Thread.sleep(2000);
