+++ src/test/java/com/qauber/sanity/ViewReport.java
@@ -29,7 +29,7 @@ public class ViewReport extends PageObjectModelResources {
 
         driver.get(Config.getBaseURL());
         Thread.sleep(10000);
-        getLogin().loginToWave(getTestCaseUser().getUsername(), getTestCaseUser().getPassword());
+        getLogin().loginToWave(testUser().getUsername(), testUser().getPassword());
         Thread.sleep(5000);
         getNavBar().clickReports();
